% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLblipWrappers.R
\name{SL.lrTE}
\alias{SL.lrTE}
\title{SL.lrTE}
\usage{
SL.lrTE(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\description{
does main terms for outcome fit on A = 0.  Uses this fit as offset in TE fitter
}
