% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLblipWrappers.R
\name{SL.xgbB}
\alias{SL.xgbB}
\title{SL.xgbB}
\usage{
SL.xgbB(Y, X, newX, family, obsWeights, id, ntrees = 1000, max_depth = 2,
  shrinkage = 0.01, minobspernode = 10, params = list(), nthread = 1,
  verbose = 0, save_period = NULL, ...)
}
\description{
boosting for interactions
}
